Reorganized league stats into game manager
