+++ lib/game_team_manager.rb
