+++ lib/team_manager.rb
@@ -2,10 +2,11 @@ require_relative 'team'
 require 'csv'
 
 class TeamManager
-  attr_reader :teams
+  attr_reader :teams, :teams_data
   def initialize(locations, stat_tracker)
     @stat_tracker = stat_tracker
     @teams = generate_teams(locations[:teams])
+    @teams_data = team_data_by_id
   end
 
   def generate_teams(location)
@@ -17,9 +18,14 @@ class TeamManager
   end
 
   def team_info(team_id)
-    team = teams.find do |team_obj|
-      team_obj.team_id == team_id
-    end
-    team.team_info
+    teams_data[team_id]
+  end
+
+  def team_data_by_id
+    @teams.map{|team| [team.team_id, team.team_info]}.to_h
+  end
+
+  def count_of_teams
+    teams_data.count
   end
 end
