+++ test/game_team_manager_test.rb
@@ -2,4 +2,5 @@ require './test/test_helper'
 require './lib/game_team_manager'
 
 class GameTeamManagerTest < Minitest::Test
+
 end
